Add swipeRefreshLayout and clean up code

+++ app/src/main/java/com/codepath/apps/simpletweets/TwitterClient.java
@@ -42,7 +42,7 @@ public class TwitterClient extends OAuthBaseClient {
      */
 
 	// DEFINE METHODS for different API endpoints here
-	public void getHomeTimeline(AsyncHttpResponseHandler handler, Long maxId, Integer sinceId) {
+	public void getHomeTimeline(AsyncHttpResponseHandler handler, Long maxId, Long sinceId) {
 		String apiUrl = getApiUrl("statuses/home_timeline.json");
 		// Can specify query string params directly or through RequestParams.
 		RequestParams params = new RequestParams();
